+++ summon-core/src/commonTest/kotlin/code/yousef/summon/modifier/AttributeModifiersTest.kt
@@ -1,0 +1,13 @@
+package code.yousef.summon.modifier
+
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class AttributeModifiersTest {
+
+    @Test
+    fun buttonTypeSetsAttribute() {
+        val modifier = Attributes.run { Modifier().buttonType(ButtonType.Submit) }
+        assertEquals("submit", modifier.attributes["type"])
+    }
+}